smdk4412: Disable samsung_extdisp driver
This is based on commit 3419d27.

The samsung_extdisp driver for GT-I9500 was affected by CVE-2015-1800 and
CVE-2015-1801. We use an older version of this driver and it is at
least affected by CVE-2015-1801. Newer kernel versions for the GT-I9500
had the driver completely removed. There is no indication that the
driver is actually needed, so let's disable it.

vulnerability disclosure and further information:
http://blog.quarkslab.com/kernel-vulnerabilities-in-the-samsung-s4.html

Change-Id: I726231929ab88fb2b41631ba3ac1522b1dfe4dca

Conflicts:
	arch/arm/configs/lineageos_t0ltekor_defconfig